+++ cloudmesh/common/parameter.py
@@ -2,10 +2,36 @@ from hostlist import expand_hostlist
 
 
 class Parameter(object):
+
     @classmethod
     def expand(cls, parameter, allow_duplicates=False, sort=False, sep=":"):
         """
         Parameter.expand("a[0-1]")  -> ["a0", "a1"]
+        Content sensitive : expansion
+        Parameter.expand("local:a0,a1")  -> ["local:a0", "local:a1"]
+        instead of
+        Parameter.expand("local:[a0,a1]")  -> ["local:a0", "local:a1"]
+
+        :param parameter:
+        :param allow_duplicates:
+        :param sort:
+        :return:
+        """
+        if parameter is None:
+            return parameter
+
+        return list(expand_hostlist(parameter,
+                                    allow_duplicates=False,
+                                    sort=False))
+    @classmethod
+    def _expand(cls, parameter, allow_duplicates=False, sort=False, sep=":"):
+        """
+        Parameter.expand("a[0-1]")  -> ["a0", "a1"]
+        Content sensitive : expansion
+        Parameter.expand("local:a0,a1")  -> ["local:a0", "local:a1"]
+        instead of
+        Parameter.expand("local:[a0,a1]")  -> ["local:a0", "local:a1"]
+
         :param parameter:
         :param allow_duplicates:
         :param sort:
@@ -18,7 +44,6 @@ class Parameter(object):
                                           allow_duplicates=False,
                                           sort=False))
 
-        print(parameters)
         results = [t.split(sep, 1) for t in parameters]
         merge = []